Optimize project1 pipeline for BeagleBone Black performance
- Switch sample arrays from float64 to float32 throughout the parser,
  halving memory usage per waveform (~800KB vs 1.6MB at 200k samples)
- Eliminate duplicate cross-correlation in compare_waveforms — reuse
  already-computed rmse/correlation/lag instead of calling compare_signals
  a second time inside compute_metrics baseline path
- Vectorize peak detection in _estimate_damping with numpy where(),
  replacing an O(n) Python loop with a single array operation (~10x faster)
- Precompute SNR convolution kernel at module level to avoid re-allocating
  np.ones(5) on every compute_metrics call
- Replace O(n²) browser DFT in renderFftChart with server-computed numpy
  FFT data (fft_freqs/fft_magnitudes now included in waveform API response)
- Fix downsampling in comparator to use np.linspace indices for evenly
  spaced points instead of stride which could overshoot by 1
- Compute comparison timestamp once per result instead of twice
- Store comparison JSON without indentation to reduce file size ~20%

+++ project1/app/ingest/parser.py
@@ -89,8 +89,8 @@ def _parse_wave_v1(filepath, max_samples=None):
         if len(sample_bytes) < read_count * 4:
             raise ValueError(f"File appears truncated ({filepath})")
 
-        # big-endian float32 → float64 for all internal processing
-        samples = np.frombuffer(sample_bytes, dtype=">f4").astype(np.float64)
+        # big-endian float32 → native float32; stays float32 throughout the pipeline
+        samples = np.frombuffer(sample_bytes, dtype=">f4").astype(np.float32)
 
     return WaveformData(
         source_id=source_id,
@@ -110,7 +110,7 @@ def parse_raw_uint8(filepath, sample_rate, source_id, units="ADC counts", max_sa
     # (each uint8 sample = 1 byte, so byte count == sample count).
     with open(filepath, "rb") as f:
         raw = np.frombuffer(f.read(max_samples), dtype=np.uint8)
-    samples = raw.astype(np.float64)
+    samples = raw.astype(np.float32)
     samples -= samples.mean()
 
     return WaveformData(
